+++ lib/pry/commands/change_prompt.rb
@@ -18,9 +18,9 @@ class Pry::Command::ChangePrompt < Pry::ClassCommand
     end
   end
 
-private
+  private
   def prompt_map
-    Pry::Prompt::MAP
+    Pry::Prompt::PROMPT_MAP
   end
   Pry::Commands.add_command(self)
 end