+++ spec/prompt_spec.rb
@@ -1,5 +1,37 @@
 require_relative 'helper'
 
+RSpec.describe Pry::Prompt do
+  let(:prompt_value) do
+    [proc{},proc{}]
+  end
+
+  before do
+    described_class.add_prompt "prompt-name", "prompt description", prompt_value
+  end
+
+  after do
+    described_class.remove_prompt "prompt-name"
+  end
+
+  describe ".add_prompt" do
+    specify "it adds a new prompt to Pry" do
+      new_prompt = described_class::PROMPT_MAP['prompt-name']
+      expect(new_prompt).to eq(description: "prompt description", value: prompt_value)
+      expect(pry_eval("list-prompts")).to include("prompt-name")
+      expect(pry_eval("list-prompts")).to include("prompt description")
+      expect(pry_eval("change-prompt prompt-name", "_pry_.prompt")).to eq(prompt_value)
+    end
+  end
+
+  describe ".remove_prompt" do
+    specify "it removes a prompt from Pry" do
+      described_class.remove_prompt 'prompt-name'
+      expect(pry_eval("list-prompts")).to_not include("prompt-name")
+      expect(pry_eval("list-prompts")).to_not include("prompt description")
+    end
+  end
+end
+
 describe "Prompts" do
   describe "one-parameter prompt proc" do
     it 'should get full config object' do