+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### HEAD
 
 * Add clear-screen command ([#1630](https://github.com/pry/pry/pull/1630))
+* Add `Pry::Prompt.add_prompt()`, `Pry::Prompt.get_prompt()`, and `Pry::Prompt.remove_prompt()`, for integrating custom prompts with Pry ([#1628](https://github.com/pry/pry/pull/1628))
 * Add text helpers for background colors ([#1624](https://github.com/pry/pry/pull/1624))
 * Fix string literal methods completion. ([#1590](https://github.com/pry/pry/pull/1590))
 * Make sure Pry::WrappedModule::Candidate#source_location returns non-nil value when `.name` has